// services/api-service/src/orders/dto/create-conditional-order.dto.d.ts
export declare class CreateConditionalOrderDto {
    marketId: string;
    tokenId: string;
    type: string;
    side: string;
    outcome: string;
    size: number;
    triggerPrice: number;
    limitPrice?: string;
    trailingPct?: string;
    expiresAt?: string;
}
//# sourceMappingURL=create-conditional-order.dto.d.ts.map